rag-app/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
pydantic>=2.0.0
# Authentication
bcrypt>=4.1.0
pyjwt>=2.8.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
email-validator>=2.0.0
# LangChain Stack
langchain>=0.1.0
langchain-community>=0.0.10
langchain-chroma>=0.1.0
langchain-openai>=0.0.5
rank_bm25>=0.2.2
# Docling
docling>=1.0.0
langchain-docling>=0.1.0
chromadb>=0.4.22
# Search Engine
meilisearch>=0.30.0
# Observability
langfuse>=2.0.0
